# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Playlist:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      videos:
        items:
          $ref: '#/definitions/domain.Video'
        type: array
    type: object
...
  domain.Video:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      qualities:
        items:
          $ref: '#/definitions/domain.VideoQuality'
        type: array
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_url:
        type: string
      views:
        type: integer
    type: object
  domain.VideoQuality:
    properties:
      created_at:
        type: string
      id:
        type: integer
      quality:
        type: string
      video_id:
        type: integer
      video_url:
        type: string
    type: object
  dto.AddVideoToPlaylistRequest:
    properties:
      video_id:
        type: integer
    type: object
...
  dto.CreatePlaylistRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdatePlaylistRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  httpdelivery.CommonError:
    properties:
      error:
        type: string
    type: object
  httpdelivery.CommonResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: analogyoutube@example.com
    name: AnalogYouTube Backend
  description: |-
    API для AnalogYouTube: пользователи, видео, чаты, донаты и остальное.
    Для закрытых запросов нажмите Authorize и вставьте: Bearer <token>.
  title: AnalogYouTube API
  version: "1.0"
paths:
...
  /api/playlists:
    post:
      consumes:
      - application/json
      description: Новый публичный плейлист.
      parameters:
      - description: Данные плейлиста
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlaylistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Создать плейлист
      tags:
      - Playlists
  /api/playlists/{id}:
    delete:
      description: Удаляет плейлист автора.
      parameters:
      - description: ID плейлиста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Удалить плейлист
      tags:
      - Playlists
    get:
      description: Плейлист и его видео.
      parameters:
      - description: ID плейлиста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Playlist'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Плейлист по ID
      tags:
      - Playlists
    put:
      consumes:
      - application/json
      description: Меняет название или описание.
      parameters:
      - description: ID плейлиста
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Playlist'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить плейлист
      tags:
      - Playlists
  /api/playlists/{id}/videos:
    post:
      consumes:
      - application/json
      description: Добавляет ролик в плейлист.
      parameters:
      - description: ID плейлиста
        in: path
        name: id
        required: true
        type: integer
      - description: ID видео
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.AddVideoToPlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Добавить видео
      tags:
      - Playlists
  /api/playlists/{id}/videos/{video_id}:
    delete:
      description: Убирает ролик из плейлиста.
      parameters:
      - description: ID плейлиста
        in: path
        name: id
        required: true
        type: integer
      - description: ID видео
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Убрать видео
      tags:
      - Playlists
...
  /api/users/{id}/playlists:
    get:
      description: Публичные плейлисты автора.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Playlist'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Плейлисты пользователя
      tags:
      - Playlists
  /api/users/{id}/subscribe:
    delete:
      description: Отмена подписки.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Отписаться
      tags:
      - Subscriptions
    post:
      description: Подписка на автора.
      parameters:
      - description: ID пользователя, на которого подписываемся
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Подписаться
      tags:
      - Subscriptions
  /api/users/{id}/subscribed:
    get:
      description: Показывает, подписан ли я на автора.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Моя подписка
      tags:
      - Subscriptions
  /api/users/{id}/subscribers/count:
    get:
      description: Сколько людей подписано на автора.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Подписчики
      tags:
      - Subscriptions
  /api/users/{id}/subscriptions/count:
    get:
      description: Сколько подписок у пользователя.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Подписки
      tags:
      - Subscriptions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"